Addition of GTKW config and change of PreventReplay.scala
Finished first implementation, fully, of all branching for functionality.
Testing finished in the case of sliding window, start of operations
and if packet is too old.

Added config for GTKW for this test to streamline testing.
Further bugfixes started, changes needed because no functionality
was present prior to this commit.

Changes in reference to commit:
d50a912db2497d9a796145384435141063f63e97