Change of PreventReplay.scala and RFC6479_NO_M_N.py
This is for debugging purposes.
Changes consist of changing the cases for the new packet, sliding window case in code and to accomodate the new test data in Python.

Python changes are now printing the memory and adjusting the memory to 8 bits wide, just like in PreventReplay.scala

Changes in reference to:
a715378786ab868fbfa630ee4bd9b726806ddbfb