Fixes of PreventReplay.scala, Python and GTKW files.
PreventReplay.scala
Fixed issues arising in the case of sliding window not updating the state correctly, too old packet not holding old valid data and fixed most of the bugs of when S is inside of the window case.
This case remains resolved in all signals but io_drop. This is because we have concurrent when() statement running with assigning the value to the found index.
Further development in this direction is halted since decoupling of the algorithm logic and the R/W hazard resolving needs to be performed.

Changes in reference to commit:
5d2c7f6f3e4b24e51f07a5273583e7dd59fb587f