Update of .scala, .py and .gtkw config
Removed d2_data, instead placed state and next_state.
This addition resolved issue of output when packet is inside of window.
Tested and verified output and all signals are verified correct as per .py test.
Removed excessive signals and comments.
What is left is to optimize sliding window case (fresh data).
For loop needs to be optimized.

Changes in reference to commit:
85db1165fcea5046d8db7cf70f9e2c1d8a6f5e76