Added new approaches to sliding window bit clearing, as per Leon's suggestions. Built and tested one approach on our Sim object and new LinuxSim object.
Implemented Linux Wireguard C implementation self test, for our component in the form of object PreventReplayLinuxSim.

LinuxSim object compares values out of the component with what the result should be. This result can be configured to what we want. Test throws a SimFailure if there is diff between its result and the components result.
Test can now be automated and let into the CI/CD/CT pipeline.

Changes in reference to commit:
ed9dc7ce