Addition of new version of PreventReplayMN.scala and GTKW config
Changed > and < conditions back to >= and <= in the block clearing part of the sliding window procedure in the algorithm.

Changed calculations for first and last so they correctly calculate the right blocks to delete. Prior to this, they always pointed (both of them) to block 0. This is incorrect because wt_ptr was not divided by n by itself before calculating the rest of the formul, but the whole formula was divided by n, always resulting in 0. This has been fixed now.

Tested using PreventReplayMNLinuxSim and test now passes.
Generated GTKW output for this test, saved configuration for future testing if needed.

Changes in reference to commit:
cc33985f7638f151b4415e51b75026ae46c72199