Addition of signal clear into PreventReplayMN.scala and RFC6479_MN.scala
Signal io.clear in PreventReplayMN.scala now clears next_state when it's set. This allows the RISCV processor to issue a clear signal to the component if we need it to.

Addition of RFC6479_MN will be a fully Scala implementation of random data generator like the Python version. This class will be used to generate large ammounts of data and the resulting values for drop. This result will be used for generating another test in PreventReplayMN.scala and validate PreventReplayMN component.

Changes in reference to commit:
07b63a91fe2c879bb145965653360da0e5e9eff8